+++ src/cursors/mod.rs
@@ -102,7 +102,9 @@ impl<'s, T: 's, E: Extras<T>> Cursor<'s, T, E> {
     }
     #[inline]
     fn blush_extras(&mut self) {
-        self.info.extras.change(self.current());
+        if !self.noeffects() {
+            self.info.extras.change(self.current(), self.pos());
+        }
     }
 
     #[inline]
@@ -143,6 +145,15 @@ impl<'s, T: 's, E: Extras<T>> CursorTrait<'s, T, E> for Cursor<'s, T, E> {
     #[inline]
     fn is_init(&self) -> bool {
         self.info.init
+    }
+    /// if `next` or `jump` can effect the [`Extras`](Extras).
+    #[inline]
+    fn noeffects(&self) -> bool {
+        self.info.noeffects
+    }
+    #[inline]
+    fn noeffects_mut(&mut self) -> &mut bool {
+        &mut self.info.noeffects
     }
     #[inline]
     fn backwards(&self) -> bool {